move class repo
