+++ app/Services/Api/GitHubApi.php
@@ -1,9 +1,9 @@
 <?php
 
 
-namespace App\Services;
+namespace App\Services\Api;
 
-use App\Services\WordSearchApiProxy;
+use App\Services\Api\Interfaces\WordSearchApiProxy;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Http\Client\RequestException;
