+++ app/Services/Api/Interfaces/WordSearchApiProxy.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\Api\Interfaces;
 
 interface WordSearchApiProxy {
